[fix] Fixed support add pages
