+++ src/pages/Group/Mail.tsx
@@ -75,74 +75,72 @@ export function MailAutoSendDialogs(props: {
   }
 
   return (
-    <div>
-      <Dialog
-        onClose={() => setOpen('')}
-        fullWidth={true}
-        aria-labelledby="customized-dialog-title"
-        open={open !== ''}
-        PaperProps={{
-          style: {
-            backgroundColor: '#2b2a2a',
-          },
-        }}
-      >
-        <DialogTitle id="mail-auto-send-dialog-title">Mail送信</DialogTitle>
-        <DialogContent dividers>
-          <Grid container spacing={3}>
-            <Grid item xs={12}>
-              <TextField
-                id="to"
-                label="To"
-                variant="outlined"
-                value={toMail}
-                onChange={(event) => {
-                  setToMail(event.target.value)
-                }}
-              />
-            </Grid>
-            <br />
-            <Grid item xs={12}>
-              <TextField
-                id="title"
-                label="Mail Subject"
-                variant="outlined"
-                inputProps={{ maxLength: 100 }}
-                fullWidth={true}
-                value={data.subject}
-                onChange={(event) => {
-                  setData({ ...data, subject: event.target.value })
-                }}
-              />
-            </Grid>
-            <br />
-            <Grid item xs={12}>
-              <TextField
-                id="message"
-                label="Mail Body"
-                multiline
-                rows={8}
-                fullWidth={true}
-                inputProps={{ maxLength: 1000 }}
-                variant="outlined"
-                value={data.content}
-                onChange={(event) => {
-                  setData({ ...data, content: event.target.value })
-                }}
-              />
-            </Grid>
+    <Dialog
+      onClose={() => setOpen('')}
+      fullWidth={true}
+      aria-labelledby="customized-dialog-title"
+      open={open !== ''}
+      PaperProps={{
+        style: {
+          backgroundColor: '#2b2a2a',
+        },
+      }}
+    >
+      <DialogTitle id="mail-auto-send-dialog-title">Mail送信</DialogTitle>
+      <DialogContent dividers>
+        <Grid container spacing={3}>
+          <Grid item xs={12}>
+            <TextField
+              id="to"
+              label="To"
+              variant="outlined"
+              value={toMail}
+              onChange={(event) => {
+                setToMail(event.target.value)
+              }}
+            />
           </Grid>
-        </DialogContent>
-        <DialogActions>
-          <Button autoFocus onClick={() => setOpen('')} color="secondary">
-            キャンセル
-          </Button>
-          <Button autoFocus onClick={() => request()} color="primary">
-            送信
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </div>
+          <br />
+          <Grid item xs={12}>
+            <TextField
+              id="title"
+              label="Mail Subject"
+              variant="outlined"
+              inputProps={{ maxLength: 100 }}
+              fullWidth={true}
+              value={data.subject}
+              onChange={(event) => {
+                setData({ ...data, subject: event.target.value })
+              }}
+            />
+          </Grid>
+          <br />
+          <Grid item xs={12}>
+            <TextField
+              id="message"
+              label="Mail Body"
+              multiline
+              rows={8}
+              fullWidth={true}
+              inputProps={{ maxLength: 1000 }}
+              variant="outlined"
+              value={data.content}
+              onChange={(event) => {
+                setData({ ...data, content: event.target.value })
+              }}
+            />
+          </Grid>
+        </Grid>
+      </DialogContent>
+      <DialogActions>
+        <Button autoFocus onClick={() => setOpen('')} color="secondary">
+          キャンセル
+        </Button>
+        <Button autoFocus onClick={() => request()} color="primary">
+          送信
+        </Button>
+      </DialogActions>
+    </Dialog>
   )
 }
 
@@ -199,88 +197,86 @@ export function MailSendDialogs(props: {
   }
 
   return (
-    <div>
-      <Dialog
-        onClose={() => setOpen(false)}
-        fullWidth={true}
-        aria-labelledby="customized-dialog-title"
-        open={open}
-        PaperProps={{
-          style: {
-            backgroundColor: '#2b2a2a',
-          },
-        }}
-      >
-        <DialogTitle id="mail-auto-send-dialog-title">Mail送信</DialogTitle>
-        <DialogContent dividers>
-          <Grid container spacing={3}>
-            <Grid item xs={12}>
-              <Select
-                labelId="mail-template-label"
-                id="mail-template-select"
-                value={processID}
-                onChange={handleChangeProcessID}
-              >
-                {template.mail_template?.map((tmp) => (
-                  <MenuItem key={'mail_template_' + tmp.id} value={tmp.id}>
-                    {tmp.id}({tmp.title})
-                  </MenuItem>
-                ))}
-              </Select>
-            </Grid>
-            <Grid item xs={12}>
-              <TextField
-                id="to"
-                label="To"
-                variant="outlined"
-                value={mails}
-                onChange={(event) => {
-                  setToMail(event.target.value)
-                }}
-              />
-            </Grid>
-            <br />
-            <Grid item xs={12}>
-              <TextField
-                id="title"
-                label="Mail Subject"
-                variant="outlined"
-                inputProps={{ maxLength: 100 }}
-                fullWidth={true}
-                value={data.subject}
-                onChange={(event) => {
-                  setData({ ...data, subject: event.target.value })
-                }}
-              />
-            </Grid>
-            <br />
-            <Grid item xs={12}>
-              <TextField
-                id="message"
-                label="Mail Body"
-                multiline
-                rows={8}
-                fullWidth={true}
-                inputProps={{ maxLength: 1000 }}
-                variant="outlined"
-                value={data.content}
-                onChange={(event) => {
-                  setData({ ...data, content: event.target.value })
-                }}
-              />
-            </Grid>
+    <Dialog
+      onClose={() => setOpen(false)}
+      fullWidth={true}
+      aria-labelledby="customized-dialog-title"
+      open={open}
+      PaperProps={{
+        style: {
+          backgroundColor: '#2b2a2a',
+        },
+      }}
+    >
+      <DialogTitle id="mail-auto-send-dialog-title">Mail送信</DialogTitle>
+      <DialogContent dividers>
+        <Grid container spacing={3}>
+          <Grid item xs={12}>
+            <Select
+              labelId="mail-template-label"
+              id="mail-template-select"
+              value={processID}
+              onChange={handleChangeProcessID}
+            >
+              {template.mail_template?.map((tmp) => (
+                <MenuItem key={'mail_template_' + tmp.id} value={tmp.id}>
+                  {tmp.id}({tmp.title})
+                </MenuItem>
+              ))}
+            </Select>
           </Grid>
-        </DialogContent>
-        <DialogActions>
-          <Button autoFocus onClick={() => setOpen(false)} color="secondary">
-            キャンセル
-          </Button>
-          <Button autoFocus onClick={() => request()} color="primary">
-            送信
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </div>
+          <Grid item xs={12}>
+            <TextField
+              id="to"
+              label="To"
+              variant="outlined"
+              value={mails}
+              onChange={(event) => {
+                setToMail(event.target.value)
+              }}
+            />
+          </Grid>
+          <br />
+          <Grid item xs={12}>
+            <TextField
+              id="title"
+              label="Mail Subject"
+              variant="outlined"
+              inputProps={{ maxLength: 100 }}
+              fullWidth={true}
+              value={data.subject}
+              onChange={(event) => {
+                setData({ ...data, subject: event.target.value })
+              }}
+            />
+          </Grid>
+          <br />
+          <Grid item xs={12}>
+            <TextField
+              id="message"
+              label="Mail Body"
+              multiline
+              rows={8}
+              fullWidth={true}
+              inputProps={{ maxLength: 1000 }}
+              variant="outlined"
+              value={data.content}
+              onChange={(event) => {
+                setData({ ...data, content: event.target.value })
+              }}
+            />
+          </Grid>
+        </Grid>
+      </DialogContent>
+      <DialogActions>
+        <Button autoFocus onClick={() => setOpen(false)} color="secondary">
+          キャンセル
+        </Button>
+        <Button autoFocus onClick={() => request()} color="primary">
+          送信
+        </Button>
+      </DialogActions>
+    </Dialog>
   )
 }
 
@@ -344,73 +340,71 @@ export function MailAutoNoticeSendDialogs(props: {
   }
 
   return (
-    <div>
-      <Dialog
-        onClose={() => setOpen(false)}
-        fullWidth={true}
-        aria-labelledby="customized-dialog-title"
-        open={open}
-        PaperProps={{
-          style: {
-            backgroundColor: '#2b2a2a',
-          },
-        }}
-      >
-        <DialogTitle id="mail-auto-send-dialog-title">Mail送信</DialogTitle>
-        <DialogContent dividers>
-          <Grid container spacing={3}>
-            <Grid item xs={12}>
-              <TextField
-                id="to"
-                label="To"
-                variant="outlined"
-                value={toMail}
-                onChange={(event) => {
-                  setToMail(event.target.value)
-                }}
-              />
-            </Grid>
-            <br />
-            <Grid item xs={12}>
-              <TextField
-                id="title"
-                label="Mail Subject"
-                variant="outlined"
-                inputProps={{ maxLength: 100 }}
-                fullWidth={true}
-                value={data.subject}
-                onChange={(event) => {
-                  setData({ ...data, subject: event.target.value })
-                }}
-              />
-            </Grid>
-            <br />
-            <Grid item xs={12}>
-              <TextField
-                id="message"
-                label="Mail Body"
-                multiline
-                rows={8}
-                fullWidth={true}
-                inputProps={{ maxLength: 1000 }}
-                variant="outlined"
-                value={data.content}
-                onChange={(event) => {
-                  setData({ ...data, content: event.target.value })
-                }}
-              />
-            </Grid>
+    <Dialog
+      onClose={() => setOpen(false)}
+      fullWidth={true}
+      aria-labelledby="customized-dialog-title"
+      open={open}
+      PaperProps={{
+        style: {
+          backgroundColor: '#2b2a2a',
+        },
+      }}
+    >
+      <DialogTitle id="mail-auto-send-dialog-title">Mail送信</DialogTitle>
+      <DialogContent dividers>
+        <Grid container spacing={3}>
+          <Grid item xs={12}>
+            <TextField
+              id="to"
+              label="To"
+              variant="outlined"
+              value={toMail}
+              onChange={(event) => {
+                setToMail(event.target.value)
+              }}
+            />
           </Grid>
-        </DialogContent>
-        <DialogActions>
-          <Button autoFocus onClick={() => setOpen(false)} color="secondary">
-            キャンセル
-          </Button>
-          <Button autoFocus onClick={() => request()} color="primary">
-            送信
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </div>
+          <br />
+          <Grid item xs={12}>
+            <TextField
+              id="title"
+              label="Mail Subject"
+              variant="outlined"
+              inputProps={{ maxLength: 100 }}
+              fullWidth={true}
+              value={data.subject}
+              onChange={(event) => {
+                setData({ ...data, subject: event.target.value })
+              }}
+            />
+          </Grid>
+          <br />
+          <Grid item xs={12}>
+            <TextField
+              id="message"
+              label="Mail Body"
+              multiline
+              rows={8}
+              fullWidth={true}
+              inputProps={{ maxLength: 1000 }}
+              variant="outlined"
+              value={data.content}
+              onChange={(event) => {
+                setData({ ...data, content: event.target.value })
+              }}
+            />
+          </Grid>
+        </Grid>
+      </DialogContent>
+      <DialogActions>
+        <Button autoFocus onClick={() => setOpen(false)} color="secondary">
+          キャンセル
+        </Button>
+        <Button autoFocus onClick={() => request()} color="primary">
+          送信
+        </Button>
+      </DialogActions>
+    </Dialog>
   )
 }
